+++ Thermostat/lib/MovingAvarageFilter/src/MovingAverageFilter.h
@@ -6,21 +6,21 @@
 class MovingAverageFilter
 {
 public:
-	MovingAverageFilter(unsigned int newDataPointsCount, int errorValue);
+	MovingAverageFilter(uint16_t newDataPointsCount, int16_t errorValue);
 
-	int process(int value);
-	int getCurrentValue();
+	int16_t process(int16_t value);
+	int16_t getCurrentValue();
 	void reset();
 
 private:
 	long sum;
-	int lastValue;
-	int values[MAX_DATA_POINTS];
-	unsigned int readingCount;
-	unsigned int currentIdx; // k stores the index of the current array read to create a circular memory through the array
-	int errorValue;
-	unsigned int totalErrors;
+	int16_t lastValue;
+	int16_t values[MAX_DATA_POINTS];
+	uint16_t readingCount;
+	uint16_t currentIdx; // k stores the index of the current array read to create a circular memory through the array
+	int16_t errorValue;
+	uint16_t totalErrors;
 protected:
-	unsigned int dataPointsCount;
+	uint16_t dataPointsCount;
 };
 #endif
